[dev] cleanup and refactor cross-project jinja-related code
