--- projects/switchconfig/uoft_switchconfig/__main__.py
@@ -1,4 +1,0 @@
-from .cli import cli
-
-if __name__ == "__main__":
-    cli()
